fix: token null issue
