+++ src/main/java/com/jasik/momsnaggingapi/domain/auth/jwt/JwtAuthFilter.java
@@ -1,14 +1,10 @@
 package com.jasik.momsnaggingapi.domain.auth.jwt;
 
-import com.jasik.momsnaggingapi.domain.user.service.UserService;
+import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.GenericFilterBean;
 
@@ -17,26 +13,16 @@ import javax.servlet.ServletException;
 import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
+@RequiredArgsConstructor
 public class JwtAuthFilter extends GenericFilterBean {
 
-    @Autowired
-    private AuthToken authToken;
-    @Autowired
-    private AuthTokenProvider authTokenProvider;
-
-    @Autowired
-    private UserService userService;
+    private final AuthTokenProvider authTokenProvider;
 
     private static final Logger logger = LoggerFactory.getLogger(JwtAuthFilter.class);
 
     public static final String AUTHORIZATION_HEADER = "Authorization";
-
-    public JwtAuthFilter(AuthToken authToken) {
-        this.authToken = authToken;
-    }
 
     private String resolveToken(HttpServletRequest request) {
         String bearerToken = request.getHeader(AUTHORIZATION_HEADER);
@@ -51,8 +37,8 @@ public class JwtAuthFilter extends GenericFilterBean {
         try {
             HttpServletRequest httpServletRequest = (HttpServletRequest) request;
             String token = resolveToken(httpServletRequest);
-            if (token != null && authToken.validateToken(token)) {
-                Authentication authentication = authToken.getAuthentication(token);
+            if (token != null && authTokenProvider.validateToken(token)) {
+                Authentication authentication = authTokenProvider.getAuthentication(token);
                 SecurityContextHolder.getContext().setAuthentication(authentication);
             }
         } catch (Exception e) {
